Move embed under resolver-map
