+++ src/resolver-map/pack.ts
@@ -1,7 +1,7 @@
 import { Packer, PackOptions, PackState } from '../types';
 import cloneDeep from 'lodash.clonedeep';
 import { embedPackOptionsWrapper } from '../utils';
-import { embed } from '../resolver/embed';
+import { embed } from './embed';
 
 export const defaultPackOptions: PackOptions = { state: {}, dependencies: {} };
